requirements.txt
# Core Azure dependencies
azure-identity>=1.15.0
openai>=1.12.0

# Vector store and embeddings
chromadb>=0.4.22

# Financial data
yfinance>=0.2.36
requests>=2.31.0  # For Polygon.io fallback

# CLI and utilities
typer>=0.9.0
rich>=13.7.0  # For beautiful CLI output

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Logging and monitoring
structlog>=24.1.0

# Social Media APIs (Phase 3)
tweepy>=4.14.0  # X (Twitter) API
praw>=7.7.1  # Reddit API
textblob>=0.17.1  # Simple sentiment analysis

# Telegram Bot (Phase 3)
python-telegram-bot>=20.7  # Telegram bot framework

# Machine Learning (Phase 4)
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0

# Azure ML (Optional - only for distributed training on Azure ML)
# If you want to train on Azure ML, install separately:
#   pip install -r ml/azure_ml/requirements-azure-ml.txt
# For local training, these are not required

# Testing (optional for Phase 1, but good to have)
pytest>=8.0.0
pytest-asyncio>=0.23.0

# UI
fastapi>=0.109.0
uvicorn>=0.27.0
